requirements.txt
PyCryptodome
pillow
tqdm
socket
os
secrets